# Optimised TopLevelModuleName, removed SourceToModule and one instance of canonicalizeAbsolutePath