--- a/Agda.cabal
+++ b/Agda.cabal
@@ -421,6 +421,7 @@ library
                     Agda.Syntax.Scope.Base
                     Agda.Syntax.Scope.Flat
                     Agda.Syntax.Scope.Monad
+                    Agda.Syntax.TopLevelModuleName
                     Agda.Syntax.Translation.AbstractToConcrete
                     Agda.Syntax.Translation.ConcreteToAbstract
                     Agda.Syntax.Translation.InternalToAbstract