--- a/src/full/Agda/Compiler/Backend.hs
+++ b/src/full/Agda/Compiler/Backend.hs
@@ -36,6 +36,7 @@ import GHC.Generics (Generic)
 
 import System.Console.GetOpt
 
+import Agda.Syntax.TopLevelModuleName
 import Agda.Syntax.Treeless
 import Agda.TypeChecking.Errors (getAllWarnings)
 -- Agda.TypeChecking.Monad.Base imports us, relying on the .hs-boot file to
@@ -83,15 +84,18 @@ data Backend' opts env menv mod def = Backend'
       --   vanilla Agda behaviour.
   , preCompile       :: opts -> TCM env
       -- ^ Called after type checking completes, but before compilation starts.
-  , postCompile      :: env -> IsMain -> Map ModuleName mod -> TCM ()
+  , postCompile      :: env -> IsMain -> Map TopLevelModuleName mod ->
+                        TCM ()
       -- ^ Called after module compilation has completed. The @IsMain@ argument
       --   is @NotMain@ if the @--no-main@ flag is present.
-  , preModule        :: env -> IsMain -> ModuleName -> Maybe FilePath -> TCM (Recompile menv mod)
+  , preModule        :: env -> IsMain -> TopLevelModuleName ->
+                        Maybe FilePath -> TCM (Recompile menv mod)
       -- ^ Called before compilation of each module. Gets the path to the
       --   @.agdai@ file to allow up-to-date checking of previously written
       --   compilation results. Should return @Skip m@ if compilation is not
       --   required. Will be @Nothing@ if only scope checking.
-  , postModule       :: env -> menv -> IsMain -> ModuleName -> [def] -> TCM mod
+  , postModule       :: env -> menv -> IsMain -> TopLevelModuleName ->
+                        [def] -> TCM mod
       -- ^ Called after all definitions of a module have been compiled.
   , compileDef       :: env -> menv -> IsMain -> Definition -> TCM def
       -- ^ Compile a single definition.
@@ -254,7 +258,9 @@ compilerMain backend isMain0 checkResult = inCompilerEnv checkResult $ do
     mods <- doCompile
         -- This inner function is called for both `Agda.Primitive` and the module in question,
         -- and all (distinct) imported modules. So avoid shadowing "isMain" or "i".
-        (\ifaceIsMain iface -> Map.singleton (iModuleName iface) <$> compileModule backend env ifaceIsMain iface)
+        (\ifaceIsMain iface ->
+          Map.singleton (iTopLevelModuleName iface) <$>
+          compileModule backend env ifaceIsMain iface)
         isMain i
     -- Note that `doCompile` calls `setInterface` for each distinct module in the graph prior to calling into
     -- `compileModule`. This last one is just to ensure it's reset to _this_ module.
@@ -263,19 +269,19 @@ compilerMain backend isMain0 checkResult = inCompilerEnv checkResult $ do
 
 compileModule :: Backend' opts env menv mod def -> env -> IsMain -> Interface -> TCM mod
 compileModule backend env isMain i = do
-  mName <- toTopLevelModuleName <$> curMName
+  mName <- curMName
   -- The interface file will only exist if performing af full type-check, vs scoping.
   -- FIXME: Expecting backends to read the timestamp of the output path of the interface
   --        file for dirtiness checking is very roundabout and heavily couples backend
   --        implementations to the filesystem as the source of cache state.
   mifile <- (Just . filePath . intFilePath =<<) <$> findInterfaceFile mName
-  r      <- preModule backend env isMain (iModuleName i) mifile
+  r      <- preModule backend env isMain (iTopLevelModuleName i) mifile
   case r of
     Skip m         -> return m
     Recompile menv -> do
       defs <- map snd . sortDefs <$> curDefs
       res  <- mapM (compileDef' backend env menv isMain <=< instantiateFull) defs
-      postModule backend env menv isMain (iModuleName i) res
+      postModule backend env menv isMain (iTopLevelModuleName i) res
 
 compileDef' :: Backend' opts env menv mod def -> env -> menv -> IsMain -> Definition -> TCM def
 compileDef' backend env menv isMain def =