--- a/src/full/Agda/Compiler/Common.hs
+++ b/src/full/Agda/Compiler/Common.hs
@@ -9,6 +9,7 @@ import qualified Data.Map as Map
 import Data.Set (Set)
 import qualified Data.Set as Set
 import qualified Data.HashMap.Strict as HMap
+import qualified Data.HashSet as HSet
 import Data.Char
 import Data.Function
 #if __GLASGOW_HASKELL__ < 804
@@ -21,12 +22,13 @@ import Control.Monad.State
 import Agda.Syntax.Common
 import qualified Agda.Syntax.Concrete.Name as C
 import Agda.Syntax.Internal as I
+import Agda.Syntax.TopLevelModuleName
 
 import Agda.Interaction.FindFile ( srcFilePath )
 import Agda.Interaction.Options
 import Agda.Interaction.Imports  ( CheckResult, crInterface, crSource, Source(..) )
 
-import Agda.TypeChecking.Monad
+import Agda.TypeChecking.Monad as TCM
 
 import Agda.Utils.FileName
 import Agda.Utils.Lens
@@ -57,11 +59,19 @@ doCompile f isMain i = do
   where
   -- The Agda.Primitive module is only loaded if the --no-load-primitives flag was not given,
   -- thus, only try to compile it if we have visited it.
-  compilePrim cont = (lift $ Map.lookup agdaPrim <$> getVisitedModules) >>= \case
-    Nothing   -> cont
-    Just prim -> mappend <$> doCompile' f NotMain (miInterface prim) <*> cont
+  compilePrim cont = do
+    agdaPrim <- lift $ do
+      agdaPrim <- TCM.topLevelModuleName agdaPrim
+      Map.lookup agdaPrim <$> getVisitedModules
+    case agdaPrim of
+      Nothing   -> cont
+      Just prim ->
+        mappend <$> doCompile' f NotMain (miInterface prim) <*> cont
     where
-    agdaPrim = C.TopLevelModuleName mempty $ "Agda" :| "Primitive" : []
+    agdaPrim = RawTopLevelModuleName
+      { rawModuleNameRange = mempty
+      , rawModuleNameParts = "Agda" :| "Primitive" : []
+      }
       -- N.B. The Range in TopLevelModuleName is ignored for Ord, so we can set it to mempty.
 
 -- This helper function is called for both `Agda.Primitive` and the module in question.
@@ -74,7 +84,7 @@ doCompile' f isMain i = do
   if alreadyDone then return mempty else do
     imps <- lift $
       map miInterface . catMaybes <$>
-        mapM (getVisitedModule . toTopLevelModuleName . fst) (iImportedModules i)
+        mapM (getVisitedModule . fst) (iImportedModules i)
     ri <- mconcat <$> mapM (doCompile' f NotMain) imps
     lift $ setInterface i
     r <- lift $ f isMain i
@@ -90,16 +100,17 @@ setInterface i = do
   -- that it doesn't suffice to replace setTCLens' with setTCLens,
   -- because the stPreImportedModules field is strict.
   stImportedModules `setTCLens'`
-    Set.fromList (map fst $ iImportedModules i)
-  stCurrentModule   `setTCLens'` Just (iModuleName i)
+    HSet.fromList (map fst (iImportedModules i))
+  stCurrentModule `setTCLens'`
+    Just (iModuleName i, iTopLevelModuleName i)
 
 curIF :: ReadTCState m => m Interface
 curIF = do
   name <- curMName
-  maybe __IMPOSSIBLE__ miInterface <$> getVisitedModule (toTopLevelModuleName name)
+  maybe __IMPOSSIBLE__ miInterface <$> getVisitedModule name
 
-curMName :: ReadTCState m => m ModuleName
-curMName = fromMaybe __IMPOSSIBLE__ <$> useTC stCurrentModule
+curMName :: ReadTCState m => m TopLevelModuleName
+curMName = maybe __IMPOSSIBLE__ snd <$> useTC stCurrentModule
 
 curDefs :: ReadTCState m => m Definitions
 curDefs = HMap.filter (not . defNoCompilation) . (^. sigDefinitions) . iSignature <$> curIF
@@ -147,9 +158,9 @@ inCompilerEnv checkResult cont = do
           Just dir -> dir
           Nothing  ->
             -- The default output directory is the project root.
-            let tm = toTopLevelModuleName $ iModuleName mainI
+            let tm = iTopLevelModuleName mainI
                 f  = srcFilePath $ srcOrigin checkedSource
-            in filePath $ C.projectRoot f tm
+            in filePath $ projectRoot f tm
     setCommandLineOptions $
       opts { optCompileDir = Just compileDir }
 
@@ -175,16 +186,18 @@ inCompilerEnv checkResult cont = do
   stTCWarnings `setTCLens` newWarnings
   return a
 
-topLevelModuleName :: ReadTCState m => ModuleName -> m ModuleName
+topLevelModuleName ::
+  ReadTCState m => ModuleName -> m TopLevelModuleName
 topLevelModuleName m = do
-  -- get the names of the visited modules
-  visited <- map (iModuleName . miInterface) . Map.elems <$>
-    getVisitedModules
+  -- Interfaces of visited modules.
+  visited <- map miInterface . Map.elems <$> getVisitedModules
   -- find the module with the longest matching prefix to m
-  let ms = sortBy (compare `on` (length . mnameToList)) $
-        filter (\ m' -> mnameToList m' `isPrefixOf` mnameToList m) visited
-  case ms of
-    (m' : _) -> return m'
+  let is = sortBy (compare `on` (length . mnameToList . iModuleName)) $
+           filter (\i -> mnameToList (iModuleName i) `isPrefixOf`
+                         mnameToList m)
+             visited
+  case is of
+    (i : _) -> return (iTopLevelModuleName i)
     -- if we did not get anything, it may be because m is a section
     -- (a module _ ), see e.g. #1866
     []       -> curMName