--- a/src/full/Agda/Compiler/MAlonzo/Compiler.hs
+++ b/src/full/Agda/Compiler/MAlonzo/Compiler.hs
@@ -54,6 +54,7 @@ import Agda.Syntax.Internal as I
 import Agda.Syntax.Internal.Names (namesIn)
 import qualified Agda.Syntax.Treeless as T
 import Agda.Syntax.Literal
+import Agda.Syntax.TopLevelModuleName
 
 import Agda.TypeChecking.Datatypes
 import Agda.TypeChecking.Primitive (getBuiltinName)
@@ -70,6 +71,7 @@ import Agda.Utils.Float
 import Agda.Utils.IO.Directory
 import Agda.Utils.Lens
 import Agda.Utils.List
+import qualified Agda.Utils.List1 as List1
 import Agda.Utils.Maybe
 import Agda.Utils.Monad
 import Agda.Utils.Pretty (prettyShow, render)
@@ -192,7 +194,7 @@ data GHCDefinition = GHCDefinition
   , ghcDefDecls      :: [HS.Decl]
   , ghcDefDefinition :: Definition
   , ghcDefMainDef    :: Maybe MainFunctionDef
-  , ghcDefImports    :: Set ModuleName
+  , ghcDefImports    :: Set TopLevelModuleName
   }
 
 --- Top-level compilation ---
@@ -324,7 +326,8 @@ ghcPreCompile flags = do
     , ghcEnvIsTCBuiltin = istcbuiltin
     }
 
-ghcPostCompile :: GHCEnv -> IsMain -> Map ModuleName GHCModule -> TCM ()
+ghcPostCompile ::
+  GHCEnv -> IsMain -> Map TopLevelModuleName GHCModule -> TCM ()
 ghcPostCompile _cenv _isMain mods = do
   -- FIXME: @curMName@ and @curIF@ are evil TCM state, but there does not appear to be
   --------- another way to retrieve the compilation root ("main" module or interaction focused).
@@ -340,7 +343,7 @@ ghcPostCompile _cenv _isMain mods = do
 ghcPreModule
   :: GHCEnv
   -> IsMain      -- ^ Are we looking at the main module?
-  -> ModuleName
+  -> TopLevelModuleName
   -> Maybe FilePath    -- ^ Path to the @.agdai@ file.
   -> TCM (Recompile GHCModuleEnv GHCModule)
                  -- ^ Could we confirm the existence of a main function?
@@ -360,7 +363,8 @@ ghcPreModule cenv isMain m mifile =
     ifileDesc = fromMaybe "(memory)" mifile
 
     noComp = do
-      reportSLn "compile.ghc" 2 . (++ " : no compilation is needed.") . prettyShow . A.mnameToConcrete =<< curMName
+      reportSLn "compile.ghc" 2 .
+        (++ " : no compilation is needed.") . prettyShow =<< curMName
       menv <- ask
       mainDefs <- ifM curIsMainModule
                          (mainFunctionDefs <$> curIF)
@@ -368,7 +372,7 @@ ghcPreModule cenv isMain m mifile =
       return . Skip $ GHCModule menv mainDefs
 
     yesComp = do
-      m   <- prettyShow . A.mnameToConcrete <$> curMName
+      m   <- prettyShow <$> curMName
       out <- curOutFile
       reportSLn "compile.ghc" 1 $ repl [m, ifileDesc, out] "Compiling <<0>> in <<1>> to <<2>>"
       asks Recompile
@@ -377,7 +381,7 @@ ghcPostModule
   :: GHCEnv
   -> GHCModuleEnv
   -> IsMain        -- ^ Are we looking at the main module?
-  -> ModuleName
+  -> TopLevelModuleName
   -> [GHCDefinition]   -- ^ Compiled module content.
   -> TCM GHCModule
 ghcPostModule _cenv menv _isMain _moduleName ghcDefs = do
@@ -425,7 +429,9 @@ ghcMayEraseType q = getHaskellPragma q <&> \case
 
 -- Compilation ------------------------------------------------------------
 
-imports :: BuiltinThings PrimFun -> Set ModuleName -> [Definition] -> [HS.ImportDecl]
+imports ::
+  BuiltinThings PrimFun -> Set TopLevelModuleName -> [Definition] ->
+  [HS.ImportDecl]
 imports builtinThings usedModules defs = hsImps ++ imps where
   hsImps :: [HS.ImportDecl]
   hsImps = [unqualRTE, decl mazRTE]
@@ -446,7 +452,7 @@ imports builtinThings usedModules defs = hsImps ++ imps where
   decl :: HS.ModuleName -> HS.ImportDecl
   decl m = HS.ImportDecl m True Nothing
 
-  mnames :: [ModuleName]
+  mnames :: [TopLevelModuleName]
   mnames = Set.elems usedModules
 
   uniq :: [HS.ModuleName] -> [HS.ModuleName]
@@ -1278,7 +1284,7 @@ callGHC = do
   opts    <- askGhcOpts
   hsmod   <- prettyPrint <$> curHsMod
   agdaMod <- curAgdaMod
-  let outputName = lastWithDefault __IMPOSSIBLE__ $ mnameToList agdaMod
+  let outputName = List1.last $ moduleNameParts agdaMod
   (mdir, fp) <- curOutFileAndDir
   let ghcopts = optGhcFlags opts
 
@@ -1293,7 +1299,7 @@ callGHC = do
 
   let overridableArgs =
         [ "-O"] ++
-        (if isMain then ["-o", mdir </> prettyShow (nameConcrete outputName)] else []) ++
+        (if isMain then ["-o", mdir </> outputName] else []) ++
         [ "-Werror"]
       otherArgs       =
         [ "-i" ++ mdir] ++