--- a/src/full/Agda/Interaction/Base.hs
+++ b/src/full/Agda/Interaction/Base.hs
@@ -112,6 +112,15 @@ data Command' a
 
 type Command = Command' IOTCM
 
+-- | IOTCM commands.
+--
+-- The commands are obtained by applying the functions to the current
+-- top-level module name, if any. Note that the top-level module name
+-- is not used by independent commands. For other commands the
+-- top-level module name should be known.
+
+type IOTCM = Maybe TopLevelModuleName -> IOTCM' Range
+
 -- | Command queues.
 
 data CommandQueue = CommandQueue
@@ -301,7 +310,6 @@ data Interaction' range
     -- ^ Exit the program.
         deriving (Show, Read, Functor, Foldable, Traversable)
 
-type IOTCM = IOTCM' Range
 data IOTCM' range
     = IOTCM
         FilePath
@@ -327,6 +335,18 @@ data Remove
 ---------------------------------------------------------
 -- Read instances
 
+-- | An 'IOTCM' parser.
+--
+-- If the parse fails, then an error message is returned.
+
+parseIOTCM ::
+  String -> Either String IOTCM
+parseIOTCM s = case listToMaybe $ reads s of
+  Just (x, "")  -> Right $ \top -> case x of
+    IOTCM f l m i -> IOTCM f l m $
+      (fmap . fmap . fmap) (\rf -> mkRangeFile (rangeFilePath rf) top) i
+  Just (_, rem) -> Left $ "not consumed: " ++ rem
+  _             -> Left $ "cannot read: " ++ s
 
 -- | The 'Parse' monad.
 --   'StateT' state holds the remaining input.
@@ -391,6 +411,14 @@ instance Read AbsolutePath where
     readsPrec = parseToReadsPrec $ do
         exact "mkAbsolute"
         fmap mkAbsolute readParse
+
+-- | This instance fills in the 'TopLevelModuleName's using 'Nothing'.
+-- Note that these occurrences of 'Nothing' are \"overwritten\" by
+-- 'parseIOTCM'.
+
+instance Read RangeFile where
+    readsPrec = parseToReadsPrec $
+      fmap (flip mkRangeFile Nothing) readParse
 
 instance Read a => Read (Position' a) where
     readsPrec = parseToReadsPrec $ do