--- a/src/full/Agda/Syntax/Position.hs-boot
+++ b/src/full/Agda/Syntax/Position.hs-boot
@@ -1,0 +1,3 @@
+module Agda.Syntax.Position where
+
+class KillRange a