… src/full/Agda/Interaction/Highlighting/Dot/Backend.hs
… src/full/Agda/Interaction/Highlighting/FromAbstract.hs
… src/full/Agda/Interaction/Highlighting/Generate.hs
… src/full/Agda/Interaction/Highlighting/HTML/Backend.hs
… src/full/Agda/Interaction/Highlighting/HTML/Base.hs
… src/full/Agda/Interaction/Highlighting/LaTeX/Backend.hs
… src/full/Agda/Interaction/Highlighting/LaTeX/Base.hs
… src/full/Agda/Interaction/Highlighting/Precise.hs
… src/full/Agda/Interaction/Imports.hs-boot
… src/full/Agda/Interaction/InteractionTop.hs
… src/full/Agda/Interaction/JSONTop.hs
… src/full/Agda/Syntax/Abstract/Name.hs
… src/full/Agda/Syntax/Common.hs
… src/full/Agda/Syntax/Concrete.hs
… src/full/Agda/Syntax/Concrete/Name.hs
… src/full/Agda/Syntax/Literal.hs
… src/full/Agda/Syntax/Parser.hs
… src/full/Agda/Syntax/Parser/Monad.hs
… src/full/Agda/Syntax/Position.hs
… src/full/Agda/Syntax/TopLevelModuleName.hs
… src/full/Agda/Syntax/TopLevelModuleName.hs-boot
… src/full/Agda/Syntax/Translation/ConcreteToAbstract.hs
… src/full/Agda/TypeChecking/Errors.hs
… src/full/Agda/TypeChecking/Monad/Base.hs
… src/full/Agda/TypeChecking/Monad/Base.hs-boot
… src/full/Agda/TypeChecking/Monad/Env.hs
… src/full/Agda/TypeChecking/Monad/Imports.hs
… src/full/Agda/TypeChecking/Monad/Options.hs
… src/full/Agda/TypeChecking/Monad/State.hs
… src/full/Agda/TypeChecking/Monad/Statistics.hs