Some of these changes are related to #6146. One of the changes fixes #6145.

Now `Range`s contain `TopLevelModuleName`s, which in turn now contain `ModuleNameHash`es to enable fast comparisons. I also replaced `String` with `Text` in `TopLevelModuleName`. I removed `SourceToModule` and `lookupModuleFromSource`, and the serialiser no longer uses `canonicalizeAbsolutePath`.

Quick testing for the standard library (with GHC 9.4.2 and `-foptimise-heavily`) suggests a speed-up of 10-20%, but as [noted before](https://github.com/agda/agda/pull/6099#issuecomment-1252039574) one should perhaps take such numbers with a grain of salt. Before:
```
 981,074,791,736 bytes allocated in the heap
  95,509,216,416 bytes copied during GC
   1,707,164,008 bytes maximum residency (166 sample(s))
       4,930,384 bytes maximum slop
            2684 MiB total memory in use (0 MB lost due to fragmentation)

                                     Tot time (elapsed)  Avg pause  Max pause
  Gen  0     231748 colls,     0 par   76.336s  76.454s     0.0003s    0.0293s
  Gen  1       166 colls,     0 par   13.247s  13.247s     0.0798s    0.8304s

  TASKS: 4 (1 bound, 3 peak workers (3 total), using -N1)

  SPARKS: 0 (0 converted, 0 overflowed, 0 dud, 0 GC'd, 0 fizzled)

  INIT    time    0.000s  (  0.000s elapsed)
  MUT     time  300.721s  (299.141s elapsed)
  GC      time   89.584s  ( 89.701s elapsed)
  EXIT    time    0.001s  (  0.008s elapsed)
  Total   time  390.307s  (388.850s elapsed)

  Alloc rate    3,262,407,786 bytes per MUT second

  Productivity  77.0% of total user, 76.9% of total elapsed


real	6m28,928s
user	6m18,086s
sys	0m12,299s
```
After (another run was faster, `real	5m19,033s`):
```
 820,385,471,096 bytes allocated in the heap
  92,492,394,712 bytes copied during GC
   1,695,477,656 bytes maximum residency (159 sample(s))
       4,837,064 bytes maximum slop
            2670 MiB total memory in use (0 MB lost due to fragmentation)

                                     Tot time (elapsed)  Avg pause  Max pause
  Gen  0     195161 colls,     0 par   72.794s  72.907s     0.0004s    0.0263s
  Gen  1       159 colls,     0 par   12.663s  12.664s     0.0796s    0.8577s

  TASKS: 4 (1 bound, 3 peak workers (3 total), using -N1)

  SPARKS: 0 (0 converted, 0 overflowed, 0 dud, 0 GC'd, 0 fizzled)

  INIT    time    0.000s  (  0.000s elapsed)
  MUT     time  253.297s  (251.793s elapsed)
  GC      time   85.457s  ( 85.571s elapsed)
  EXIT    time    0.001s  (  0.006s elapsed)
  Total   time  338.756s  (337.370s elapsed)

  Alloc rate    3,238,833,943 bytes per MUT second

  Productivity  74.8% of total user, 74.6% of total elapsed


real	5m37,449s
user	5m35,123s
sys	0m3,711s
```
Note that the `sys` time is significantly lower.